s3: read presign expiresIn as i64 so every out-of-range value reports the SigV4 bound
With the i32 read, values above i32::MAX were rejected by the integer
width check first and the error named the i32 bounds instead of
1..604800. Reading as i64 (like partSize) routes them all through the
same range error.

+++ src/runtime/webcore/S3File.rs
@@ -753,15 +753,15 @@ pub(crate) fn get_presign_url_from(
                     }
                 };
             }
-            if let Some(expires_) = options.get_optional_int::<i32>(global, "expiresIn")? {
-                const MAX_EXPIRES: i32 =
-                    bun_s3_signing::SignQueryOptions::MAX_EXPIRES_SECONDS as i32;
+            if let Some(expires_) = options.get_optional_int::<i64>(global, "expiresIn")? {
+                const MAX_EXPIRES: i64 =
+                    bun_s3_signing::SignQueryOptions::MAX_EXPIRES_SECONDS as i64;
                 if !(1..=MAX_EXPIRES).contains(&expires_) {
                     return Err(global.throw_range_error(
-                        i64::from(expires_),
+                        expires_,
                         RangeErrorOptions {
                             min: 1,
-                            max: i64::from(MAX_EXPIRES),
+                            max: MAX_EXPIRES,
                             field_name: b"expiresIn",
                             ..Default::default()
                         },
